ActivityPub API: KeyPackage management

+++ handler/activitypub_user/outbox_keyPackages.go
@@ -16,8 +16,7 @@ func init() {
 	outboxRouter.Add(vocab.ActivityTypeDelete, vocab.ObjectTypeKeyPackage, outbox_DeleteKeyPackage)
 }
 
-// This funciton handles ActivityPub "Accept/Follow" activities, meaning that
-// it is called with a remote server accepts our follow request.
+// Create a new KeyPackage record from the ActivityPub API
 func outbox_CreateKeyPackage(context Context, activity streams.Document) error {
 
 	const location = "handler.activitypub_user.outbox_CreateKeyPackage"
@@ -29,6 +28,7 @@ func outbox_CreateKeyPackage(context Context, activity streams.Document) error {
 		return derp.ForbiddenError(location, "KeyPackage must be attributed to the actor")
 	}
 
+	// Populate the new KeyPackage
 	keyPackageService := context.factory.KeyPackage()
 	keyPackage := model.NewKeyPackage()
 
@@ -38,21 +38,16 @@ func outbox_CreateKeyPackage(context Context, activity streams.Document) error {
 	keyPackage.Content = object.Content()
 	keyPackage.Generator = object.Generator().ID()
 
+	// Save the KeyPackage to the database
 	if err := keyPackageService.Save(context.session, &keyPackage, "Created via ActivityPub API"); err != nil {
 		return derp.Wrap(err, location, "Unable to save KeyPackage")
 	}
 
+	// In Vino veritas
 	return nil
 }
 
-func outbox_AddKeyPackage(context Context, activity streams.Document) error {
-	return outbox_SetKeyPackageVisibility(context, activity, true)
-}
-
-func outbox_RemoveKeyPackage(context Context, activity streams.Document) error {
-	return outbox_SetKeyPackageVisibility(context, activity, true)
-}
-
+// Locate and delete the KeyPackage referenced in the ActivityPub request
 func outbox_DeleteKeyPackage(context Context, activity streams.Document) error {
 
 	const location = "handler.activitypub_user.outbox_DeleteKeyPackage"
@@ -65,6 +60,7 @@ func outbox_DeleteKeyPackage(context Context, activity streams.Document) error {
 		return derp.ForbiddenError(location, "KeyPackage must be owned by this actor")
 	}
 
+	// Load the KeyPackage
 	keyPackageService := context.factory.KeyPackage()
 	keyPackage := model.NewKeyPackage()
 
@@ -72,17 +68,36 @@ func outbox_DeleteKeyPackage(context Context, activity streams.Document) error {
 		return derp.Wrap(err, location, "Unable to load KeyPackage", "url", object.ID())
 	}
 
+	// RULE: The actor must own the keyPackage
+	if keyPackage.UserID != context.user.UserID {
+		return derp.ForbiddenError(location, "KeyPackage must be owned by this actor")
+	}
+
+	// Delete the KeyPackage
 	if err := keyPackageService.Delete(context.session, &keyPackage, "Deleted via ActivityPub API"); err != nil {
 		return derp.Wrap(err, location, "Unable to delete KeyPackage")
 	}
 
+	// Win.
 	return nil
 }
 
+// Add a KeyPackage to the user's collection (make it public)
+func outbox_AddKeyPackage(context Context, activity streams.Document) error {
+	return outbox_SetKeyPackageVisibility(context, activity, true)
+}
+
+// Remove a KeyPackage from the user's collection (make it private)
+func outbox_RemoveKeyPackage(context Context, activity streams.Document) error {
+	return outbox_SetKeyPackageVisibility(context, activity, true)
+}
+
+// Set the visibility of a KeyPackage in the user's collection
 func outbox_SetKeyPackageVisibility(context Context, activity streams.Document, isPublic bool) error {
 
 	const location = "handler.activitypub_user.outbox_AddKeyPackage"
 
+	// Collect values from the activity
 	actor := activity.Actor()
 	object := activity.Object()
 	target := activity.Target()
@@ -97,6 +112,7 @@ func outbox_SetKeyPackageVisibility(context Context, activity streams.Document, 
 		return derp.ForbiddenError(location, "KeyPackage must be owned by this actor")
 	}
 
+	// Load the KeyPackage
 	keyPackageService := context.factory.KeyPackage()
 	keyPackage := model.NewKeyPackage()
 
@@ -104,11 +120,14 @@ func outbox_SetKeyPackageVisibility(context Context, activity streams.Document, 
 		return derp.Wrap(err, location, "Unable to load KeyPackage", "url", object.ID())
 	}
 
+	// Set the visibility
 	keyPackage.IsPublic = isPublic
 
+	// Save the KeyPackage to the database
 	if err := keyPackageService.Save(context.session, &keyPackage, "Published via ActivityPub API"); err != nil {
 		return derp.Wrap(err, location, "Unable to save KeyPackage")
 	}
 
+	// Yup.
 	return nil
 }
